// src/shared/api/models/UpdateTaskBody.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type UpdateTaskBody = {
  /**
   * @description ID задачи
   * @type number
   */
  id: number
  /**
   * @description Заголовок задачи
   * @type string
   */
  title: string
  /**
   * @description Дата окончания задачи
   * @type string
   */
  deadline_date?: string | null
  /**
   * @description Дата выполнения задачи
   * @type string
   */
  done_date?: string | null
  /**
   * @description Примечание задачи
   * @type string
   */
  note?: string | null
}